Sellman problem with stable baseline.
Pending to check whether it converges or not. There is an issue when trying to render it, now the line is commented to skip the error. Next step: use a action wrap to modifiy  the available actions. It would be more autonomous to train with all action and giving it a vector as state of the previous points